+++ src/Pim/Bundle/FilterBundle/Resources/public/js/datafilter/filters-manager.js
@@ -31,6 +31,57 @@ define(
                 '</select>'
             ),
 
+            enableFilters: function(filters) {
+                if (_.isEmpty(filters)) {
+                    return this;
+                }
+                var optionsSelectors = [];
+
+                _.each(filters, function(filter) {
+                    filter.render();
+                    filter.enable();
+                    optionsSelectors.push('option[value="' + filter.name + '"]:not(:selected)');
+                }, this);
+
+                var options = this.$(this.filterSelector).find(optionsSelectors.join(','));
+                if (options.length) {
+                    options.attr('selected', true);
+                }
+
+                if (optionsSelectors.length) {
+                    this.selectWidget.multiselect('refresh');
+                }
+
+                return this;
+            },
+
+            render: function () {
+                this.$el.empty();
+                var fragment = document.createDocumentFragment();
+
+                _.each(this.filters, function(filter) {
+                    if (filter.enabled) {
+                        filter.render();
+                    } else {
+                        filter.hide();
+                    }
+
+                    fragment.appendChild(filter.$el.get(0));
+                }, this);
+
+                this.trigger("rendered");
+
+                if (_.isEmpty(this.filters)) {
+                    this.$el.hide();
+                } else {
+                    this.$el.append(this.addButtonTemplate({filters: this.filters}));
+                    this.$el.append(fragment);
+                    this._initializeSelectWidget();
+                }
+
+                return this;
+            },
+
             _initializeSelectWidget: function() {
                 FiltersManager.prototype._initializeSelectWidget.apply(this, arguments);